fix(stylexswc/css-parser): hold every dimension and number at double width
The remaining numeric CSS types stored single precision where the
compiler they must agree with holds a JavaScript number, so an authored
value was rounded before anything else happened to it. Angle, time,
frequency, resolution, percentage, bare number, flex fraction, alpha and
the `calc()` dimension now hold a double, and the three conversions that
happen at print time -- milliseconds to seconds, Hz to KHz, and the
`calc()` operand chain -- compute at that width.

One expectation moved, and it moved toward the official compiler rather
than away from it. A percentage was stored as a fraction and multiplied
back up at construction, so `7%` printed as `7.000000000000001%`. The
spec for this work predicted that string and called it parity; checking
it against `@stylexjs/babel-plugin@0.19.0` says otherwise. Upstream's
tokenizer runs `parseFloat` over the authored text and stores `7`, and
upstream emits `7%`. The percentage token carries the authored percent
here now, and the division survives only where a fraction is genuinely
wanted -- an alpha, a filter argument, a scale factor -- which is one
division instead of a round trip. It also settles a disagreement inside
this crate, where a percentage built from a token printed `0.5%` and the
same percentage built from its own type printed `50%`.

The colour channels still hold single precision and are widened in their
own commits; the two places that meet them are marked as bridges.

Every expectation added here was confirmed against the official compiler
run over the same value. Six of the new edge cases failed on first run
and all six were the test being wrong rather than the parser: the nearest
double to `0.12345678901234567` really is `...566`, an escaped `\70x` is
a valid `px`, an out-of-range alpha is carried through rather than
refused, and an unclosed `calc(` is tolerated rather than rejected. Each
is now pinned as the behaviour it actually is.

+++ crates/stylex-css-parser/src/css_types/angle.rs
@@ -15,13 +15,13 @@ pub const ANGLE_UNITS: &[&str] = &["deg", "grad", "rad", "turn"];
 /// CSS Angle value with unit
 #[derive(Debug, Clone, PartialEq)]
 pub struct Angle {
-  pub value: f32,
+  pub value: f64,
   pub unit: String, // "deg", "grad", "rad", or "turn"
 }
 
 impl Angle {
   /// Create a new Angle value
-  pub fn new(value: f32, unit: impl Into<String>) -> Self {
+  pub fn new(value: f64, unit: impl Into<String>) -> Self {
     Self {
       value,
       unit: unit.into(),
@@ -58,7 +58,7 @@ impl Angle {
   /// branch reachable from coverage tests.
   pub(crate) fn extract_dimension_token(token: SimpleToken) -> Angle {
     if let SimpleToken::Dimension { value, unit } = token {
-      Angle::new(value as f32, unit)
+      Angle::new(value, unit)
     } else {
       stylex_unreachable!()
     }
