+++ crates/stylex-css-parser/src/css_types/resolution.rs
@@ -15,13 +15,13 @@ pub const RESOLUTION_UNITS: &[&str] = &["dpi", "dpcm", "dppx"];
 /// CSS Resolution value with unit
 #[derive(Debug, Clone, PartialEq)]
 pub struct Resolution {
-  pub value: f32,
+  pub value: f64,
   pub unit: String, // "dpi", "dpcm", or "dppx"
 }
 
 impl Resolution {
   /// Create a new Resolution value
-  pub fn new(value: f32, unit: impl Into<String>) -> Self {
+  pub fn new(value: f64, unit: impl Into<String>) -> Self {
     Self {
       value,
       unit: unit.into(),
@@ -58,7 +58,7 @@ impl Resolution {
   /// branch reachable from coverage tests.
   pub(crate) fn extract_dimension_token(token: SimpleToken) -> Resolution {
     if let SimpleToken::Dimension { value, unit } = token {
-      Resolution::new(value as f32, unit)
+      Resolution::new(value, unit)
     } else {
       stylex_unreachable!()
     }
