+++ crates/stylex-css-parser/src/tests/css_types/angle_percentage_tests.rs
@@ -23,10 +23,10 @@ fn test_angle_percentage_display() {
   let percentage = AnglePercentage::Percentage(Percentage::new(25.0));
   assert_eq!(percentage.to_string(), "25%");
 
-  let rad_angle = AnglePercentage::Angle(Angle::new(std::f32::consts::PI, "rad".to_string()));
+  let rad_angle = AnglePercentage::Angle(Angle::new(std::f64::consts::PI, "rad".to_string()));
   assert_eq!(
     rad_angle.to_string(),
-    format!("{}rad", std::f32::consts::PI)
+    format!("{}rad", std::f64::consts::PI)
   );
 
   let zero_angle = AnglePercentage::Angle(Angle::new(0.0, "deg".to_string()));
